requirements.txt
requests
unicodecsv
lxml
simplejson
mwclient
pyyaml
